app/requirements.txt
opencv-python>=4.8
ultralytics>=8.2
pyyaml>=6.0
requests>=2.31
numpy>=1.24
